requirements.txt
# Core dependencies
google-generativeai>=0.3.0
python-dotenv>=1.0.0

# Data Sources
yfinance>=0.2.28
requests>=2.31.0
polygon-api-client>=1.12.3  # Polygon.io API client
websocket-client>=1.6.1    # For WebSocket streaming data
redis>=5.0.0               # For caching frequently queried data

# Data Processing & Analysis
pandas>=2.0.3
numpy>=1.24.3
matplotlib>=3.7.2
statsmodels>=0.14.0
prophet>=1.1.4
# tensorflow>=2.13.0 # Install manually if needed
scikit-learn>=1.3.0
sympy>=1.12.0
scipy>=1.10.1  # For portfolio optimization

# Visualization & Output
matplotlib>=3.7.2
plotly>=5.16.1
kaleido>=0.2.1  # For plotly static image export
openpyxl>=3.1.2  # For Excel export
pdfkit>=1.0.0  # For PDF generation
seaborn>=0.12.2  # For enhanced visualizations

# Web & Data Fetching
beautifulsoup4>=4.12.2
requests>=2.31.0
python-dotenv>=1.0.0
tornado>=6.3.3  # For alert system
twilio>=8.2.2  # For alert notifications

# AI & NLP Tools
transformers>=4.30.2
torch>=2.0.1
aif360>=0.5.0  # For bias mitigation
nltk>=3.8.1  # For VADER sentiment analysis
finbert-embedding>=0.1.4  # For financial sentiment analysis
twitter-api-client>=0.1.3  # For social media sentiment analysis
textblob>=0.17.1  # For simple sentiment analysis fallback
huggingface-hub>=0.16.4  # For downloading pre-trained models

# Interactive Visualization
plotly>=5.14.0
dash>=2.9.0
streamlit>=1.22.0
bokeh>=3.1.0

# Database & Storage
sqlite3-tools>=0.1  # For multi-user support and persistance

# Portfolio Management & Backtesting
PyPortfolioOpt>=1.5.4  # For portfolio optimization (alternative to scipy implementation)
backtrader>=1.9.76.123  # For advanced backtesting (optional)
